+++ src/main/java/nl/sugcube/dirtyarrows/ability/Airship.java
@@ -5,6 +5,9 @@ import nl.sugcube.dirtyarrows.DirtyArrows;
 import org.bukkit.entity.Player;
 import org.bukkit.entity.Projectile;
 
+/**
+ * Handler for the airship bow: makes sure to constantly keep flying.
+ */
 public class Airship implements Runnable {
 
 public static DirtyArrows plugin;