+++ src/main/java/nl/sugcube/dirtyarrows/ability/TreeCut.java
@@ -1,13 +1,16 @@
 package nl.sugcube.dirtyarrows.ability;
 
-import java.util.Random;
-
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 
+import java.util.Random;
+
+/**
+ * Utility class with just static helpers.
+ */
 public class TreeCut {
 	
 	public static boolean returnValue;