+++ src/main/kotlin/nl/sugcube/dirtyarrows/DirtyArrows.kt
@@ -114,7 +114,6 @@ class DirtyArrows : JavaPlugin() {
         registerEvents(pjl, plugin)
         registerEvents(pdl, plugin)
         registerEvents(iron, plugin)
-        registerEvents(curse, plugin)
         registerEvents(frozenListener, plugin)
         registerEvents(anvilListener, plugin)
     }